tratando exceção delete

+++ src/main/java/com/luizfelipe/one_project/service/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.luizfelipe.one_project.entities.User;
 import com.luizfelipe.one_project.repositories.UserRepository;
+import com.luizfelipe.one_project.service.exceptions.DataBaseException;
 import com.luizfelipe.one_project.service.exceptions.ResourceNotFoundException;
 
 @Service
@@ -33,7 +35,16 @@ public class UserService {
 	}
 	
 	public void delete (Long id) {
+
+		if (!userRepository.existsById(id)) {
+		    throw new ResourceNotFoundException(id);
+		}
+		try {
 		userRepository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
 	
 	public User update (Long id, User obj) {
